Refactor playlist creation and sorting logic; enhance error handling and add utility functions for weighted results
